requirements.txt
streamlit>=1.28.0
gitpython>=3.1.40
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
markdown>=3.5.0
beautifulsoup4>=4.12.0
pygments>=2.16.0
plotly>=5.17.0
